commands and queries moved from domain to application
