+++ Services/UserContactInformation/UserContactInformation.Application/Features/Contact/Commands/RemoveContactCommandHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace UserContactInformation.Application.Features.Contact.Commands
+{
+    internal class RemoveContactCommandHandler
+    {
+    }
+}
