+++ Services/UserContactInformation/UserContactInformation.Application/Features/Person/Commands/AddPersonCommand.cs
@@ -1,0 +1,6 @@
+﻿namespace UserContactInformation.Application.Features.Person.Commands
+{
+    internal class AddPersonCommand
+    {
+    }
+}
