+++ Services/UserContactInformation/UserContactInformation.Application/Features/Person/Commands/RemovePersonCommand.cs
@@ -1,0 +1,6 @@
+﻿namespace UserContactInformation.Application.Features.Person.Commands
+{
+    internal class RemovePersonCommand
+    {
+    }
+}
