+++ Services/UserContactInformation/UserContactInformation.Application/Features/Person/Commands/RemovePersonCommandHandler.cs
@@ -1,0 +1,6 @@
+﻿namespace UserContactInformation.Application.Features.Person.Commands
+{
+    internal class RemovePersonCommandHandler
+    {
+    }
+}
