+++ Services/UserContactInformation/UserContactInformation.Application/Features/Person/Queries/ListPersonsQuery.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace UserContactInformation.Domain.Features.Report.Queries
+namespace UserContactInformation.Application.Features.Person.Queries
 {
-    internal class ListReportsQuery
+    internal class ListPersonsQuery
     {
     }
 }
