+++ Services/UserContactInformation/UserContactInformation.Application/Features/Report/Queries/GetReportDetailByIdQuery.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace UserContactInformation.Domain.Features.Report.Queries
+namespace UserContactInformation.Application.Features.Report.Queries
 {
-    internal class ListReportsQueryHandler
+    internal class GetReportDetailByIdQuery
     {
     }
 }
