+++ Services/UserContactInformation/UserContactInformation.Application/Features/Report/Queries/ListReportsQueryHandler.cs
@@ -4,9 +4,9 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace UserContactInformation.Domain.Features.Report.Commands
+namespace UserContactInformation.Application.Features.Report.Queries
 {
-    internal class ProduceReportCommand
+    internal class ListReportsQueryHandler
     {
     }
 }
