--- Services/UserContactInformation/UserContactInformation.Domain/Features/Contact/Commands/AddContactCommand.cs
@@ -1,6 +1,0 @@
-﻿namespace UserContactInformation.Domain.Features.Contact.Commands
-{
-    internal class AddContactCommand
-    {
-    }
-}
