--- Services/UserContactInformation/UserContactInformation.Domain/Features/Person/Commands/RemovePersonCommand.cs
@@ -1,6 +1,0 @@
-﻿namespace UserContactInformation.Domain.Features.Person.Commands
-{
-    internal class RemovePersonCommand
-    {
-    }
-}
