--- Services/UserContactInformation/UserContactInformation.Domain/Features/Person/Commands/RemovePersonCommandHandler.cs
@@ -1,6 +1,0 @@
-﻿namespace UserContactInformation.Domain.Features.Person.Commands
-{
-    internal class RemovePersonCommandHandler
-    {
-    }
-}
